placesview: implement new GtkPlacesView widget
GtkPlacesView is the next step related to the ongoing effort
to modernize Gtk+ components and the whole GNOME stack as
well.

The new widget monitors permanent devices (like hard drives'
partitions) and manages network connections. It enables mounting
devices and connecting to networks, as well as storing a list
of recent servers.