placessidebar: add Other Locations item
Places sidebar is a widget that enabled the user to
select XDG directories, bookmarks and mounted network
locations, as well as manages permanent and removable
devices.

The new design that aims to look less clutered makes
the sidebar display only removable devices, as well as
mounted networks, bookmarks and XDG directories, and
delegates the management of permanent devices such as
hard drive partitions to GtkPlacesView, a newly
introduced widget for this specific purpose.

To delegate it, add an "Other Locations..." item to
notify when the permanent devices manager is required.
Besides that, don't show these fixes devices on the
sidebar itself, as they not supposed to be handled by
the sidebar anymore..